feat: extract EditorAdvancedModal
