+++ src/app/dashboard/[subdomain]/editor/page-component.tsx
@@ -7,6 +7,7 @@ import dynamic from "next/dynamic"
 import { useParams, useRouter, useSearchParams } from "next/navigation"
 import {
   ChangeEvent,
+  FC,
   memo,
   useCallback,
   useEffect,
@@ -30,7 +31,6 @@ import { EditorToolbar } from "~/components/ui/EditorToolbar"
 import { FieldLabel } from "~/components/ui/FieldLabel"
 import { ImageUploader } from "~/components/ui/ImageUploader"
 import { Input } from "~/components/ui/Input"
-import { Modal } from "~/components/ui/Modal"
 import { ModalContentProps, useModalStack } from "~/components/ui/ModalStack"
 import { Switch } from "~/components/ui/Switch"
 import { TagInput } from "~/components/ui/TagInput"
@@ -41,7 +41,7 @@ import {
   Values,
   initialEditorState,
   useEditorState,
-} from "~/hooks/useEdtiorState"
+} from "~/hooks/useEditorState"
 import { useGetState } from "~/hooks/useGetState"
 import { useIsMobileLayout } from "~/hooks/useMobileLayout"
 import { useBeforeMounted } from "~/hooks/useSyncOnce"
@@ -519,9 +519,7 @@ export default function SubdomainEditor() {
     )
   }, [draftKey, subdomain, site.data?.characterId])
 
-  const [isAdvancedOptionsOpen, setIsAdvancedOptionsOpen] =
-    useState<boolean>(false)
-
+  const presentAdvancedModal = useEditorAdvancedModal({ isPost })
   const extraProperties = (
     <EditorExtraProperties
       defaultSlug={defaultSlug}
@@ -529,7 +527,7 @@ export default function SubdomainEditor() {
       isPost={isPost}
       subdomain={subdomain}
       userTags={userTags.data?.list || []}
-      openAdvancedOptions={() => setIsAdvancedOptionsOpen(true)}
+      openAdvancedOptions={presentAdvancedModal}
     />
   )
 
@@ -715,19 +713,13 @@ export default function SubdomainEditor() {
                   isPost={isPost}
                   userTags={userTags.data?.list || []}
                   subdomain={subdomain}
-                  openAdvancedOptions={() => setIsAdvancedOptionsOpen(true)}
+                  openAdvancedOptions={presentAdvancedModal}
                 />
               )}
             </div>
           </>
         )}
       </DashboardMain>
-      <EditorAdvancedOptions
-        updateValue={updateValue}
-        isAdvancedOptionsOpen={isAdvancedOptionsOpen}
-        setIsAdvancedOptionsOpen={setIsAdvancedOptionsOpen}
-        isPost={isPost}
-      />
     </>
   )
 }
@@ -874,96 +866,98 @@ const EditorExtraProperties = memo(
 
 EditorExtraProperties.displayName = "EditorExtraProperties"
 
-const EditorAdvancedOptions = memo(
-  ({
-    updateValue,
-    isAdvancedOptionsOpen,
-    setIsAdvancedOptionsOpen,
-    isPost,
-  }: {
-    updateValue: <K extends keyof Values>(key: K, value: Values[K]) => void
+const useEditorAdvancedModal = ({ isPost }: { isPost: boolean }) => {
+  const { t } = useTranslation("dashboard")
 
-    isAdvancedOptionsOpen: boolean
-    setIsAdvancedOptionsOpen: (state: boolean) => void
-    isPost: boolean
-  }) => {
-    const values = useEditorState(
-      (state) => pick(state, ["disableAISummary", "publishedAt"]),
-      shallow,
-    )
-    const { t } = useTranslation("dashboard")
+  const { present } = useModalStack()
 
-    return (
-      <Modal
-        open={isAdvancedOptionsOpen}
-        setOpen={setIsAdvancedOptionsOpen}
-        title={t("Advanced Settings")}
-      >
-        <div className="p-5 space-y-5">
-          <div>
-            <label className="form-label">
-              {t("Disable AI-generated summary")}
-            </label>
-            <Switch
-              label=""
-              checked={values.disableAISummary}
-              setChecked={(state) => updateValue("disableAISummary", state)}
-            />
-          </div>
-          <div>
-            <label className="form-label" htmlFor="publishAt">
-              {t("Publish at")}
-            </label>
-            <DateInput
-              className=""
-              allowDeselect
-              clearable
-              valueFormat="YYYY-MM-DD, h:mm a"
-              name="publishAt"
-              id="publishAt"
-              value={
-                values.publishedAt ? new Date(values.publishedAt) : undefined
-              }
-              onChange={(value: Date | null) => {
-                if (value) {
-                  updateValue("publishedAt", value.toISOString())
-                } else {
-                  updateValue("publishedAt", "")
-                }
-              }}
-              styles={{
-                input: {
-                  borderRadius: "0.5rem",
-                  borderColor: "var(--border-color)",
-                  height: "2.5rem",
-                  "&:focus-within": {
-                    borderColor: "var(--theme-color)",
-                  },
-                },
-              }}
-            />
-            <div className="text-xs text-gray-400 mt-1">
-              {t(
-                `This ${
-                  isPost ? "post" : "page"
-                } will be accessible from this time. Leave blank to use the current time.`,
-              )}
-            </div>
-            {values.publishedAt > new Date().toISOString() && (
-              <div className="text-xs mt-1 text-orange-500">
-                {t(
-                  "The post is currently not public as its publication date has been scheduled for a future time.",
-                )}
-              </div>
+  return () => {
+    present({
+      title: t("Advanced Settings"),
+      content: () => <EditorAdvancedModal isPost={isPost} />,
+    })
+  }
+}
+
+const EditorAdvancedModal: FC<{
+  isPost: boolean
+}> = ({ isPost }) => {
+  const { t } = useTranslation("dashboard")
+
+  const values = useEditorState(
+    (state) => pick(state, ["disableAISummary", "publishedAt"]),
+    shallow,
+  )
+  const updateValue = useEditorState.setState
+  return (
+    <div className="p-5 space-y-5">
+      <div>
+        <label className="form-label">
+          {t("Disable AI-generated summary")}
+        </label>
+        <Switch
+          label=""
+          checked={values.disableAISummary}
+          // setChecked={(state) => updateValue("disableAISummary", state)}
+          setChecked={(state) => {
+            updateValue({
+              disableAISummary: state,
+            })
+          }}
+        />
+      </div>
+      <div>
+        <label className="form-label" htmlFor="publishAt">
+          {t("Publish at")}
+        </label>
+        <DateInput
+          className="[&_input]:bg-slate-50 [&_input]:text-black/90"
+          allowDeselect
+          clearable
+          valueFormat="YYYY-MM-DD, h:mm a"
+          name="publishAt"
+          id="publishAt"
+          value={values.publishedAt ? new Date(values.publishedAt) : undefined}
+          onChange={(value: Date | null) => {
+            if (value) {
+              updateValue({
+                publishedAt: value.toISOString(),
+              })
+            } else {
+              updateValue({
+                publishedAt: "",
+              })
+            }
+          }}
+          styles={{
+            input: {
+              borderRadius: "0.5rem",
+              borderColor: "var(--border-color)",
+              height: "2.5rem",
+              "&:focus-within": {
+                borderColor: "var(--theme-color)",
+              },
+            },
+          }}
+        />
+        <div className="text-xs text-gray-400 mt-1">
+          {t(
+            `This ${
+              isPost ? "post" : "page"
+            } will be accessible from this time. Leave blank to use the current time.`,
+          )}
+        </div>
+        {values.publishedAt > new Date().toISOString() && (
+          <div className="text-xs mt-1 text-orange-500">
+            {t(
+              "The post is currently not public as its publication date has been scheduled for a future time.",
             )}
           </div>
-        </div>
-      </Modal>
-    )
-  },
-)
-
-EditorAdvancedOptions.displayName = "EditorAdvancedOptions"
+        )}
+      </div>
+    </div>
+  )
+}
 
 const PublishedModal = ({
   dismiss,
