+++ src/components/ui/Modal.tsx
@@ -78,7 +78,7 @@ export const Modal = forwardRef<HTMLDivElement, ModalProps>(
               {/* The actual dialog panel  */}
               <Dialog.Panel
                 className={cn(
-                  `mx-auto rounded-lg bg-white w-full shadow-modal max-h-full overflow-y-auto flex flex-col`,
+                  `mx-auto rounded-lg bg-white w-full shadow-modal max-h-full flex flex-col`,
                   size === "md"
                     ? `max-w-md`
                     : size === "lg"
@@ -94,7 +94,7 @@ export const Modal = forwardRef<HTMLDivElement, ModalProps>(
                       {title}
                     </span>
                     <span
-                      className="absolute right-1 w-7 h-7 text-xl cursor-pointer bg-white flex items-center justify-center"
+                      className="absolute right-4 w-7 h-7 text-xl cursor-pointer bg-white flex items-center justify-center"
                       onClick={() => setOpen(false)}
                     >
                       <i className="icon-[mingcute--close-line] inline-block" />
